requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv

datapizza-ai
datapizza-ai-clients-google
datapizza-ai-tools-sqldatabase

# Per frontend CRM
chainlit
asyncpg
requests
